Refactor Without Loan components
- WithoutLoan will be the component that does all logic and pass the data to the child components
