+++ src/lib/withoutLoan/GetLoan.svelte
@@ -1,64 +1,23 @@
 <script lang="ts">
-  import { BalanceDto, LoanAssetRequest, getLoanIntent } from 'pluto-loans-sdk';
-  import { Circle } from 'svelte-loading-spinners';
+  import { loanAmount } from '../verifyAccount/store';
+  import { error } from './store';
 
-  import server from '../stellar/server';
-  import { borrower } from '../verifyAccount/store';
-  import { error, loanXdr, signedXdr } from './store';
-
-  let amount: number;
-  let isLoading = false;
-
-  function getAccountXlmBalance() {
-    for (const asset of $borrower.balance) {
-      if (asset.code === 'native') {
-        return asset.balance;
-      }
-    }
-
-    throw new Error();
-  }
-
-  async function handleGetLoan() {
-    try {
-      const asset = new LoanAssetRequest(true);
-      const entryBalance = new BalanceDto(asset, `${amount}`);
-      isLoading = true;
-      clearStores();
-      $loanXdr = await getLoanIntent(server, $borrower.publicKey, entryBalance);
-    } catch (e) {
-      if (e instanceof Error) {
-        const parsedError = JSON.parse(e.message);
-        $error = `Error status ${parsedError.status}: ${parsedError.detail}`;
-      }
-    } finally {
-      isLoading = false;
-    }
-  }
-
-  function clearStores() {
-    $loanXdr = '';
-    $error = '';
-    $signedXdr = '';
-  }
+  export let balance: string;
+  export let handleGetLoan: () => void;
 </script>
 
 <h4>Get Loan intent</h4>
 
 <div class="loan-intent-container">
-  <p>{`XLM in account ${getAccountXlmBalance()}`}</p>
+  <p>{`XLM in account ${balance}`}</p>
 
   <label>
     <p>Amount in XLM:</p>
-    <input type="number" bind:value={amount} />
+    <input type="number" bind:value={$loanAmount} />
   </label>
 
-  <button on:click={handleGetLoan} disabled={amount ? false : true}>
-    {#if isLoading}
-      <Circle size="20" color="black" />
-    {:else}
-      <p>Get loan</p>
-    {/if}
+  <button on:click={handleGetLoan} disabled={$loanAmount ? false : true}>
+    <p>Get loan</p>
   </button>
 
   {#if $error}
